refactor out draggable slider
